Add standard template files and metadata
Adding these files to git will mean that if we want to remove them from
git in the future, we will need to manually ensure that they stay on the
server, otherwise things will break. They have been added to git as
there's no other (known) way to do 'migrations' but for files.